requirements.txt
flask==3.1.1
flask-login==0.6.3
GeoAlchemy2==0.17.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
shapely==2.1.1
SQLAlchemy==2.0.41
